// src/lib/enums.ts
export enum themeColor {
    /**
     * 默认（灰）
     */
    default = 'default',
    /**
     * 主色（蓝）
     */
    primary = 'primary',
    /**
     * 次色（浅蓝）
     */
    secondary = 'secondary',
    /**
     * 成功（绿）
     */
    success = 'success',
    /**
     * 警告（橙）
     */
    warning = 'warning',
    /**
     * 高危（红）
     */
    danger = 'danger'
}